Move custom collection to type directory
