+++ lib/src/types/synchronous_propagator.dart
