+++ lib/src/types/traverseable_collection.dart
