+++ lib/src/types/types.dart
@@ -2,3 +2,5 @@
 /// SPDX-License-Identifier: BSD-3-Clause
 
 export 'redriven_monitor_set.dart';
+export 'synchronous_propagator.dart';
+export 'traverseable_collection.dart';
